+++ apps/backend/src/app/views/templates/FormDeactivatedNotification.tsx
@@ -1,6 +1,7 @@
 import { Body, Head, Html, Text } from '@react-email/components'
-import { FormDeactivatedNotificationHtmlData } from '../../services/mail/mail.types'
 import { SUPPORT_FORM_LINK } from 'formsg-shared/constants'
+
+import { FormDeactivatedNotificationHtmlData } from '../../services/mail/mail.types'
 
 export const FormDeactivatedNotification = ({
   formTitle,
@@ -13,10 +14,11 @@ export const FormDeactivatedNotification = ({
       <Body>
         <Text>Dear form admins(s),</Text>
         <Text>
-          Your <b>{formTitle}</b> {' '} (<a href={formLink}>{formLink}</a>)
-          has been closed to new responses because it has exceeded the number of free SMSes.
-          Please fill out this form (<a href={SUPPORT_FORM_LINK}>go.gov.sg/formsg-support</a>)
-          to reach out to the FormSG support team.
+          Your <b>{formTitle}</b> (<a href={formLink}>{formLink}</a>) has been
+          closed to new responses because it has exceeded the number of free
+          SMSes. Please fill out this form (
+          <a href={SUPPORT_FORM_LINK}>go.gov.sg/formsg-support</a>) to reach out
+          to the FormSG support team.
         </Text>
 
         <Text>The {appName} Support Team</Text>